support real paths in coverage sensor for C++/CLI - move getInputFileIfInProject to CxxReportSensor - introduce executeImpl (template method pattern)
solution from @jmecosta #1688 with review comments from @ivangalkin
